+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+0.36 (2024-10-19)
+===============
+* [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/26)
+* Implement react lazy loading in react app.js
+* Add loader to MetricsItem when fetching data
+* Create baseUrl function and refactor api calls
+* Create manifest json file
+
 0.35 (2024-10-19)
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/25)